Refactor link batch capture modal into own component
