+++ perma_web/frontend/components/CreateLink.vue
@@ -8,6 +8,12 @@ import LinkCount from './LinkCount.vue';
 import FolderSelect from './FolderSelect.vue';
 import { useStorage } from '@vueuse/core'
 import { useToast } from '../lib/notifications';
+import CreateLinkBatch from './CreateLinkBatch.vue';
+
+const childDialogRef = ref('')
+function handleChildDialogOpen() {
+    childDialogRef.value.handleDialogOpen();
+}
 
 const userLink = ref('')
 const userLinkGUID = ref('')
@@ -138,21 +144,6 @@ onBeforeUnmount(() => {
     clearInterval(progressInterval)
 });
 
-const dialogRef = ref('')
-const handleDialogOpen = () => {
-    dialogRef.value.showModal();
-}
-
-const handleDialogClose = () => {
-    dialogRef.value.close();
-}
-
-const handleDialogClick = (e) => {
-    if (e.target.classList.contains('c-dialog')) {
-        dialogRef.value.close();
-    }
-}
-
 </script>
 
 <template>
@@ -184,9 +175,8 @@ const handleDialogClick = (e) => {
                             <ProgressBar v-if="globalStore.captureStatus === 'isCapturing'"
                                 :progress="userLinkProgressBar" />
                         </button>
-                        <p>or <button @click.prevent="handleDialogOpen" class="c-button c-button--link">create
+                        <p>or <button @click.prevent="handleChildDialogOpen" class="c-button c-button--link">create
                                 multiple links</button></p>
-
                     </div>
                     <LinkCount v-if="globalStore.userTypes.includes('individual')" />
                     <FolderSelect v-if="!globalStore.userTypes.includes('individual')" option="customSelect" />
@@ -213,43 +203,5 @@ const handleDialogClick = (e) => {
         </div>
     </div>
 
-    <dialog class="c-dialog" ref="dialogRef" @click="handleDialogClick">
-        <div class="modal-dialog modal-content">
-            <div class="modal-header">
-                <button type="button" class="close" data-dismiss="modal" @click.prevent="handleDialogClose">
-                    <span aria-hidden="true">&times;</span>
-                    <span class="sr-only" id="loading">Close</span>
-                </button>
-                <h3 id="batch-modal-title" class="modal-title">Create A Link Batch</h3>
-            </div>
-            <!-- <div class="spinner _hide">
-                <span class="sr-only" id="loading" tabindex="-1">Loading</span>
-            </div> -->
-            <div class="modal-body">
-                <div id="batch-create-input">
-                    <div class="form-group">
-                        <FolderSelect />
-                    </div>
-                    <div class="form-group">
-                        <textarea aria-label="Paste your URLs here (one URL per line)"
-                            placeholder="Paste your URLs here (one URL per line)"></textarea>
-                    </div>
-                    <div class="form-buttons">
-                        <button id="start-batch" class="btn" disabled="disabled">Create Links</button>
-                        <button class="btn cancel" @click.prevent="handleDialogClose">Cancel</button>
-                    </div>
-                </div>
-
-                <div id="batch-details-wrapper">
-                    <p id="batch-progress-report"></p>
-                    <div id="batch-details" aria-describedby="batch-progress-report"></div>
-                    <div class="form-buttons">
-                        <button class="btn cancel" data-dismiss="modal">Exit</button>
-                        <a href="#" id="export-csv" class="btn _hide" @click.prevent="handleDialogClose">Export list as
-                            CSV</a>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </dialog>
+    <CreateLinkBatch ref="childDialogRef" />
 </template>
